+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/systems/lift/LiftConstants.java
@@ -1,0 +1,17 @@
+package org.firstinspires.ftc.teamcode.synchropather.systems.lift;
+
+/**
+ * A static class used by Lift Movements to reference important kinematic and algorithmic tuning values.
+ */
+public class LiftConstants {
+
+    /**
+     *  Max lift velocity in ticks/s.
+     */
+    public static final double MAX_VELOCITY = 2000;
+    /**
+     *  Max lift acceleration in ticks/s^2.
+     */
+    public static final double MAX_ACCELERATION = 2000;
+
+}
