+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/systems/rotation/LinearRotation.java
@@ -1,6 +1,6 @@
 package org.firstinspires.ftc.teamcode.synchropather.systems.rotation;
 
-import org.firstinspires.ftc.teamcode.synchropather.DriveConstants;
+import org.firstinspires.ftc.teamcode.DriveConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.MovementType;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.TimeSpan;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.calculators.StretchedDisplacementCalculator;
