+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/systems/rotation/RotationConstants.java
@@ -1,0 +1,8 @@
+package org.firstinspires.ftc.teamcode.synchropather.systems.rotation;
+
+/**
+ * A static class used by Rotation Movements to reference important kinematic and algorithmic tuning values.
+ */
+public class RotationConstants {
+
+}
