+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/systems/rotation/RotationPlan.java
@@ -3,11 +3,9 @@ package org.firstinspires.ftc.teamcode.synchropather.systems.rotation;
 import com.arcrobotics.ftclib.geometry.Pose2d;
 
 import org.firstinspires.ftc.teamcode.RobotSystem;
-import org.firstinspires.ftc.teamcode.synchropather.DriveConstants;
 import org.firstinspires.ftc.teamcode.synchropather.systems.MovementType;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.superclasses.Movement;
 import org.firstinspires.ftc.teamcode.synchropather.systems.__util__.superclasses.Plan;
-import org.firstinspires.ftc.teamcode.synchropather.systems.translation.TranslationState;
 
 import java.util.ArrayList;
