+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/synchropather/systems/translation/TranslationConstants.java
@@ -1,0 +1,13 @@
+package org.firstinspires.ftc.teamcode.synchropather.systems.translation;
+
+/**
+ * A static class used by Translation Movements to reference important kinematic and algorithmic tuning values.
+ */
+public class TranslationConstants {
+
+    /**
+     *  Used for differentiating and integrating spline paths, between 0 and 1 (lower = more calculations, more detail).
+     */
+    public static final double delta_t = 0.005;
+
+}
